+++ client/private/js/src/systemApps/appStore/addNewApp/addNewApp.controller.js
@@ -13,15 +13,16 @@ class AddNewAppController {
     this.view = new AddNewAppView();
   }
 
-  openWindow() {
+  async openWindow() {
     if (this.model.isOpen)
       return;
 
+    const thisUserCodePenUsername = await this.model.getThisUserCodepenUsername();
     this.model.processId = Utils.randomString( 4 );
     windowManager.openNewWindowCustom(
       this.model.processId,
       'Add New App',
-      AddNewAppTemplates.content,
+      AddNewAppTemplates.content( thisUserCodePenUsername ),
       false, null,
       '30%', '65%'
     );
@@ -31,6 +32,7 @@ class AddNewAppController {
       e.preventDefault();
       try {
         const res = await this.model.addNewApp( this.view.getFormData( this.model.processId ) );
+        // TODO: (FRONTEND) Prompt the user if not successful.
         this.view.closeWindowBtnElem( this.model.processId ).click();
         console.debug( res );